--- app/cdap/components/RecipeSteps/iconStore.tsx
@@ -1,263 +1,0 @@
-/*
- * Copyright © 2022 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-import React from 'react';
-
-export const DownloadIcon = (
-  <svg width="23" height="23" viewBox="0 0 23 23" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M21.625 15V19.5C21.625 20.0967 21.3879 20.669 20.966 21.091C20.544 21.5129 19.9717 21.75 19.375 21.75H3.625C3.02826 21.75 2.45597 21.5129 2.03401 21.091C1.61205 20.669 1.375 20.0967 1.375 19.5V15"
-      stroke="#828282"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M5.875 9.375L11.5 15L17.125 9.375"
-      stroke="#828282"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M11.5 15V1.5"
-      stroke="#828282"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-);
-
-export const KebabIcon = (
-  <svg width="5" height="22" viewBox="0 0 5 22" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M2.8125 11.8125C3.46834 11.8125 4 11.2808 4 10.625C4 9.96916 3.46834 9.4375 2.8125 9.4375C2.15666 9.4375 1.625 9.96916 1.625 10.625C1.625 11.2808 2.15666 11.8125 2.8125 11.8125Z"
-      fill="#828282"
-      stroke="#828282"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M2.8125 3.5C3.46834 3.5 4 2.96834 4 2.3125C4 1.65666 3.46834 1.125 2.8125 1.125C2.15666 1.125 1.625 1.65666 1.625 2.3125C1.625 2.96834 2.15666 3.5 2.8125 3.5Z"
-      fill="#828282"
-      stroke="#828282"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M2.8125 20.125C3.46834 20.125 4 19.5933 4 18.9375C4 18.2817 3.46834 17.75 2.8125 17.75C2.15666 17.75 1.625 18.2817 1.625 18.9375C1.625 19.5933 2.15666 20.125 2.8125 20.125Z"
-      fill="#828282"
-      stroke="#828282"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-);
-
-export const InfoGraphicData = (
-  <svg
-    width="163"
-    height="245"
-    viewBox="0 0 163 245"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M149.072 84.6985V93.325C149.072 95.5303 148.343 97.6737 146.998 99.4218L92.6737 170.05C91.3292 171.798 90.6002 173.941 90.6002 176.147V210.369C90.6002 213.125 89.4629 215.759 87.4569 217.648L76.027 228.415C69.6463 234.425 59.1703 229.902 59.1703 221.136V176.152C59.1703 173.943 58.4391 171.797 57.0908 170.048L2.65956 99.424C1.31128 97.6747 0.580078 95.5282 0.580078 93.3195V84.6985C0.580078 82.3855 2.00332 80.5352 3.78238 80.5352H145.751C147.53 80.5352 149.072 82.3855 149.072 84.6985Z"
-      fill="url(#paint0_linear_1939_24942)"
-    />
-    <g filter="url(#filter0_d_1939_24942)">
-      <rect
-        x="81"
-        y="55.0508"
-        width="61.904"
-        height="14.3205"
-        rx="7.16025"
-        fill="url(#paint1_linear_1939_24942)"
-      />
-      <path
-        d="M81.0001 79.6778C81.0001 75.7233 84.2058 72.5176 88.1603 72.5176H135.744C139.698 72.5176 142.904 75.7233 142.904 79.6778C142.904 83.6323 139.698 86.8381 135.744 86.8381H88.1603C84.2058 86.8381 81.0001 83.6323 81.0001 79.6778Z"
-        fill="url(#paint2_linear_1939_24942)"
-      />
-      <ellipse cx="134.334" cy="62.1435" rx="1.09734" ry="1.09267" fill="#E94235" />
-      <ellipse cx="134.334" cy="79.579" rx="1.09734" ry="1.09267" fill="#E94235" />
-      <ellipse cx="130.676" cy="62.1435" rx="1.09734" ry="1.09267" fill="#2DA94F" />
-      <ellipse cx="130.676" cy="79.579" rx="1.09734" ry="1.09267" fill="#2DA94F" />
-      <ellipse cx="127.018" cy="62.1435" rx="1.09734" ry="1.09267" fill="#3E82F7" />
-      <ellipse cx="127.018" cy="79.579" rx="1.09734" ry="1.09267" fill="#3E82F7" />
-      <g filter="url(#filter1_d_1939_24942)">
-        <rect x="92.7452" y="16.5" width="3.45025" height="1.47238" fill="white" />
-      </g>
-    </g>
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M43.6264 89.4206C43.5021 88.7383 43.586 88.0345 43.8672 87.4001C44.135 86.778 44.5796 86.2475 45.1463 85.8737C45.713 85.5 46.3771 85.2995 47.0569 85.2968H47.247C48.3674 85.2968 49.442 84.8536 50.2343 84.0647C51.0266 83.2758 51.4717 82.2058 51.4717 81.0901C51.4717 79.9744 51.0266 78.9044 50.2343 78.1155C49.442 77.3266 48.3674 76.8834 47.247 76.8834H46.8879C46.2081 76.8806 45.544 76.6801 44.9773 76.3064C44.4106 75.9327 43.966 75.4021 43.6982 74.78V74.6117C43.417 73.9773 43.3332 73.2736 43.4574 72.5913C43.5816 71.9089 43.9083 71.2793 44.3953 70.7836L44.522 70.6574C44.9148 70.2667 45.2264 69.8028 45.439 69.2921C45.6517 68.7814 45.7611 68.234 45.7611 67.6812C45.7611 67.1283 45.6517 66.5809 45.439 66.0702C45.2264 65.5595 44.9148 65.0956 44.522 64.7049C44.1297 64.3138 43.6637 64.0035 43.1509 63.7918C42.638 63.5801 42.0882 63.4711 41.5331 63.4711C40.9779 63.4711 40.4281 63.5801 39.9152 63.7918C39.4024 64.0035 38.9364 64.3138 38.5441 64.7049L38.4173 64.8311C37.9195 65.316 37.2872 65.6413 36.602 65.765C35.9167 65.8887 35.21 65.8052 34.5729 65.5252C33.9481 65.2586 33.4153 64.8159 33.0399 64.2516C32.6646 63.6872 32.4632 63.026 32.4605 62.3491V62.1598C32.4605 61.0442 32.0154 59.9742 31.2231 59.1852C30.4308 58.3963 29.3563 57.9531 28.2358 57.9531C27.1153 57.9531 26.0408 58.3963 25.2485 59.1852C24.4562 59.9742 24.0111 61.0442 24.0111 62.1598V62.5174C24.0084 63.1942 23.807 63.8555 23.4317 64.4198C23.0564 64.9841 22.5235 65.4269 21.8988 65.6935H21.7298C21.0926 65.9735 20.3859 66.057 19.7007 65.9333C19.0154 65.8096 18.3831 65.4843 17.8853 64.9994L17.7585 64.8732C17.3662 64.482 16.9002 64.1718 16.3874 63.9601C15.8745 63.7484 15.3248 63.6394 14.7696 63.6394C14.2144 63.6394 13.6646 63.7484 13.1518 63.9601C12.6389 64.1718 12.1729 64.482 11.7806 64.8732C11.3878 65.2639 11.0762 65.7278 10.8636 66.2385C10.651 66.7492 10.5415 67.2966 10.5415 67.8494C10.5415 68.4023 10.651 68.9497 10.8636 69.4603C11.0762 69.971 11.3878 70.435 11.7806 70.8257L11.9073 70.9519C12.3943 71.4476 12.721 72.0772 12.8452 72.7595C12.9695 73.4418 12.8856 74.1456 12.6044 74.78C12.363 75.4336 11.9297 75.9997 11.3609 76.4049C10.792 76.81 10.1139 77.0353 9.41475 77.0516H9.22464C8.10418 77.0516 7.02961 77.4948 6.23733 78.2837C5.44504 79.0726 4.99994 80.1426 4.99994 81.2583C4.99994 82.374 5.44504 83.444 6.23733 84.2329C7.02961 85.0218 8.10418 85.4651 9.22464 85.4651H9.58374C10.2635 85.4678 10.9276 85.6683 11.4943 86.042C12.061 86.4157 12.5056 86.9463 12.7734 87.5684C13.0546 88.2028 13.1385 88.9066 13.0142 89.5889C12.89 90.2712 12.5633 90.9008 12.0763 91.3965L11.9496 91.5227C11.5568 91.9134 11.2452 92.3774 11.0326 92.8881C10.82 93.3987 10.7105 93.9461 10.7105 94.499C10.7105 95.0518 10.82 95.5992 11.0326 96.1099C11.2452 96.6206 11.5568 97.0845 11.9496 97.4752C12.3419 97.8664 12.8079 98.1766 13.3207 98.3883C13.8336 98.6 14.3834 98.709 14.9386 98.709C15.4937 98.709 16.0435 98.6 16.5564 98.3883C17.0692 98.1766 17.5352 97.8664 17.9275 97.4752L18.0543 97.349C18.5521 96.8641 19.1844 96.5388 19.8696 96.4151C20.5549 96.2914 21.2616 96.3749 21.8988 96.6549C22.5551 96.8953 23.1237 97.3267 23.5306 97.8932C23.9374 98.4596 24.1637 99.1348 24.1801 99.831V100.02C24.1801 101.136 24.6252 102.206 25.4175 102.995C26.2098 103.784 27.2843 104.227 28.4048 104.227C29.5253 104.227 30.5998 103.784 31.3921 102.995C32.1844 102.206 32.6295 101.136 32.6295 100.02V99.6627C32.6322 98.9859 32.8336 98.3246 33.2089 97.7603C33.5842 97.196 34.1171 96.7533 34.7418 96.4866C35.379 96.2067 36.0857 96.1231 36.7709 96.2469C37.4562 96.3706 38.0885 96.6959 38.5863 97.1808L38.7131 97.307C39.1054 97.6981 39.5714 98.0084 40.0842 98.2201C40.5971 98.4318 41.1469 98.5407 41.702 98.5407C42.2572 98.5407 42.807 98.4318 43.3199 98.2201C43.8327 98.0084 44.2987 97.6981 44.691 97.307C45.0838 96.9163 45.3954 96.4523 45.608 95.9416C45.8206 95.4309 45.9301 94.8835 45.9301 94.3307C45.9301 93.7779 45.8206 93.2305 45.608 92.7198C45.3954 92.2091 45.0838 91.7451 44.691 91.3545L44.5643 91.2283C44.0773 90.7325 43.7506 90.1029 43.6264 89.4206ZM34.5728 81.0896C34.5728 84.5746 31.7356 87.3997 28.2357 87.3997C24.7358 87.3997 21.8986 84.5746 21.8986 81.0896C21.8986 77.6047 24.7358 74.7795 28.2357 74.7795C31.7356 74.7795 34.5728 77.6047 34.5728 81.0896Z"
-      fill="url(#paint3_linear_1939_24942)"
-    />
-    <path
-      opacity="0.4"
-      d="M28.2357 87.3994C31.7356 87.3994 34.5728 84.5743 34.5728 81.0894C34.5728 77.6044 31.7356 74.7793 28.2357 74.7793C24.7359 74.7793 21.8987 77.6044 21.8987 81.0894C21.8987 84.5743 24.7359 87.3994 28.2357 87.3994Z"
-      fill="white"
-    />
-    <defs>
-      <filter
-        id="filter0_d_1939_24942"
-        x="61"
-        y="0.5"
-        width="101.904"
-        height="110.338"
-        filterUnits="userSpaceOnUse"
-        color-interpolation-filters="sRGB"
-      >
-        <feFlood flood-opacity="0" result="BackgroundImageFix" />
-        <feColorMatrix
-          in="SourceAlpha"
-          type="matrix"
-          values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-          result="hardAlpha"
-        />
-        <feOffset dy="4" />
-        <feGaussianBlur stdDeviation="10" />
-        <feComposite in2="hardAlpha" operator="out" />
-        <feColorMatrix
-          type="matrix"
-          values="0 0 0 0 0.533333 0 0 0 0 0.533333 0 0 0 0 0.533333 0 0 0 0.2 0"
-        />
-        <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_1939_24942" />
-        <feBlend
-          mode="normal"
-          in="SourceGraphic"
-          in2="effect1_dropShadow_1939_24942"
-          result="shape"
-        />
-      </filter>
-      <filter
-        id="filter1_d_1939_24942"
-        x="88.7452"
-        y="14.5"
-        width="11.4503"
-        height="9.47266"
-        filterUnits="userSpaceOnUse"
-        color-interpolation-filters="sRGB"
-      >
-        <feFlood flood-opacity="0" result="BackgroundImageFix" />
-        <feColorMatrix
-          in="SourceAlpha"
-          type="matrix"
-          values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-          result="hardAlpha"
-        />
-        <feOffset dy="2" />
-        <feGaussianBlur stdDeviation="2" />
-        <feComposite in2="hardAlpha" operator="out" />
-        <feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.2 0" />
-        <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_1939_24942" />
-        <feBlend
-          mode="normal"
-          in="SourceGraphic"
-          in2="effect1_dropShadow_1939_24942"
-          result="shape"
-        />
-      </filter>
-      <linearGradient
-        id="paint0_linear_1939_24942"
-        x1="0.580078"
-        y1="151.093"
-        x2="103.108"
-        y2="151.729"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop offset="0.0364583" stop-color="#F1F3F5" />
-        <stop offset="0.34375" stop-color="#FAFBFB" stop-opacity="0.78" />
-        <stop offset="0.630208" stop-color="#FAFAFA" />
-        <stop offset="1" stop-color="#F1F3F5" />
-      </linearGradient>
-      <linearGradient
-        id="paint1_linear_1939_24942"
-        x1="81"
-        y1="61.221"
-        x2="123.707"
-        y2="62.4839"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop offset="0.0364583" stop-color="#F1F3F5" />
-        <stop offset="0.25" stop-color="#FAFBFB" stop-opacity="0.78" />
-        <stop offset="0.390625" stop-color="#F6F7F8" />
-      </linearGradient>
-      <linearGradient
-        id="paint2_linear_1939_24942"
-        x1="81.0001"
-        y1="78.6878"
-        x2="123.707"
-        y2="79.9507"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop offset="0.0364583" stop-color="#F1F3F5" />
-        <stop offset="0.25" stop-color="#FAFBFB" stop-opacity="0.78" />
-        <stop offset="0.390625" stop-color="#F6F7F8" />
-      </linearGradient>
-      <linearGradient
-        id="paint3_linear_1939_24942"
-        x1="49.6502"
-        y1="101.015"
-        x2="8.52547"
-        y2="103.862"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stop-color="#F1F1F2" />
-        <stop offset="0.571494" stop-color="#E9E9EA" />
-        <stop offset="1" stop-color="#E5E6E6" />
-      </linearGradient>
-    </defs>
-  </svg>
-);
-
-export const DeleteIcon = (
-  <svg width="20" height="22" viewBox="0 0 20 22" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M1 5H3H19"
-      stroke="#757575"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M6 5V3C6 2.46957 6.21071 1.96086 6.58579 1.58579C6.96086 1.21071 7.46957 1 8 1H12C12.5304 1 13.0391 1.21071 13.4142 1.58579C13.7893 1.96086 14 2.46957 14 3V5M17 5V19C17 19.5304 16.7893 20.0391 16.4142 20.4142C16.0391 20.7893 15.5304 21 15 21H5C4.46957 21 3.96086 20.7893 3.58579 20.4142C3.21071 20.0391 3 19.5304 3 19V5H17Z"
-      stroke="#757575"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M8 10V16"
-      stroke="#757575"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M12 10V16"
-      stroke="#757575"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-);
